Добавил каманду, чтобы смотреть матчи Домы
