+++ src/commands/damir_stats/damir_stats.ts
@@ -1,0 +1,8 @@
+import { TContext } from "../../config";
+import { getUserMatchesListMessage } from "../../features";
+
+export const damir_stats = async (ctx: TContext) => {
+  if (!process.env.DAMIR_ID) throw new Error("Не найдена переменная DAMIR_ID");
+
+  await getUserMatchesListMessage({ ctx, userId: process.env.DAMIR_ID });
+};
