+++ src/commands/damir_stats/index.ts
@@ -1,0 +1,1 @@
+export { damir_stats } from "./damir_stats.ts";
