+++ src/commands/index.ts
@@ -1,1 +1,2 @@
 export { demapk_stats } from "./demapk_stats";
+export { damir_stats } from "./damir_stats";
