+++ src/features/getUserMatchesListMessage/index.ts
@@ -1,0 +1,1 @@
+export { getUserMatchesListMessage } from "./getUserMatchesListMessage.ts";
