+++ src/features/getUserMatchesListMessage/interfaces.ts
@@ -1,0 +1,7 @@
+import { TContext } from "../../config";
+
+export interface IGetUserMatchesListMessage {
+  ctx: TContext;
+  userId: string;
+  date?: string;
+}
